# platform/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "DirectSaleIn": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "payment": {
            "$ref": "#/components/schemas/PaymentIn"
          },
          "product_code": {
            "maxLength": 50,
            "pattern": "^[A-Z][A-Z0-9_]*$",
            "title": "Product Code",
            "type": "string"
          }
        },
        "required": [
          "product_code",
          "payment"
        ],
        "title": "DirectSaleIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RenewalIn": {
        "properties": {
          "payment": {
            "$ref": "#/components/schemas/PaymentIn"
          },
          "product_code": {
            "maxLength": 50,
            "pattern": "^[A-Z][A-Z0-9_]*$",
            "title": "Product Code",
            "type": "string"
          }
        },
        "required": [
          "product_code",
          "payment"
        ],
        "title": "RenewalIn",
        "type": "object"
      },
...
      "SubscriptionOut": {
        "properties": {
          "covers_until": {
            "format": "date-time",
            "title": "Covers Until",
            "type": "string"
          },
          "days_left": {
            "title": "Days Left",
            "type": "integer"
          },
          "last_payment": {
            "additionalProperties": true,
            "title": "Last Payment",
            "type": "object"
          },
          "organization_id": {
            "format": "uuid",
            "title": "Organization Id",
            "type": "string"
          },
          "organization_name": {
            "title": "Organization Name",
            "type": "string"
          },
          "product_code": {
            "title": "Product Code",
            "type": "string"
          },
          "product_name": {
            "title": "Product Name",
            "type": "string"
          },
          "status": {
            "enum": [
              "ACTIVE",
              "EXPIRING",
              "EXPIRED"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "organization_id",
          "organization_name",
          "product_code",
          "product_name",
          "covers_until",
          "status",
          "days_left",
          "last_payment"
        ],
        "title": "SubscriptionOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "BENILAB Digital360 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/organizations/{organization_id}/renewals": {
      "post": {
        "description": "Encaissement d'un renouvellement : la période repart de l'échéance (ou d'aujourd'hui).",
        "operationId": "admin_renew_subscription_api_v1_admin_organizations__organization_id__renewals_post",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Organization Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenewalIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Renew Subscription",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/admin/organizations/{organization_id}/sales": {
      "post": {
        "description": "Vente encaissée hors plateforme pour un client qui n'a pas fait de demande.",
        "operationId": "admin_direct_sale_api_v1_admin_organizations__organization_id__sales_post",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Organization Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectSaleIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffPurchaseRequestOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Direct Sale",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/purchase-requests": {
      "get": {
        "operationId": "admin_list_purchase_requests_api_v1_admin_purchase_requests_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PurchaseRequestStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffPurchaseRequestPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin List Purchase Requests",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/purchase-requests/{request_id}": {
      "patch": {
        "operationId": "admin_update_purchase_request_api_v1_admin_purchase_requests__request_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseRequestUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffPurchaseRequestOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Update Purchase Request",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/purchase-requests/{request_id}/void-duplicate": {
      "post": {
        "description": "Annule un paiement en double (ADMIN, FINANCE). 409 NOT_A_DUPLICATE s'il est le seul.",
        "operationId": "admin_void_duplicate_payment_api_v1_admin_purchase_requests__request_id__void_duplicate_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoidDuplicateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffPurchaseRequestOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Void Duplicate Payment",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/admin/subscriptions": {
      "get": {
        "description": "Abonnements payés hors ligne, de l'échéance la plus proche à la plus lointaine.",
        "operationId": "admin_list_subscriptions_api_v1_admin_subscriptions_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "ACTIVE",
                    "EXPIRING",
                    "EXPIRED"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin List Subscriptions",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/orgs/{org_id}/billing": {
      "get": {
        "description": "Abonnements en cours (échéance) et historique des paiements, avec leur numéro de reçu.",
        "operationId": "get_client_billing_api_v1_orgs__org_id__billing_get",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientBillingOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Client Billing",
        "tags": [
          "billing"
        ]
      }
    },
    "/api/v1/orgs/{org_id}/catalog": {
      "get": {
        "description": "Offres aux prix que l'entreprise paiera : devise imposée, pas de paramètre `currency`.\n\nPrix Europe dès qu'un numéro de l'entreprise est hors d'Afrique (catalogue v4).",
        "operationId": "get_organization_catalog_api_v1_orgs__org_id__catalog_get",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Organization Catalog",
        "tags": [
          "billing"
        ]
      }
    },
    "/api/v1/orgs/{org_id}/checkouts": {
      "post": {
        "description": "Lance un paiement en ligne (achat ou renouvellement) : rediriger vers `checkout_url`.",
        "operationId": "start_checkout_api_v1_orgs__org_id__checkouts_post",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Checkout",
        "tags": [
          "billing"
        ]
      }
    },
    "/api/v1/orgs/{org_id}/checkouts/{checkout_id}": {
      "get": {
        "description": "Retour de la page de paiement : vérifie auprès de Cartflox et active l'offre si payé.",
        "operationId": "get_checkout_api_v1_orgs__org_id__checkouts__checkout_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "checkout_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Checkout Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Checkout",
        "tags": [
          "billing"
        ]
      }
    },
...
    "/api/v1/orgs/{org_id}/purchase-requests": {
      "get": {
        "operationId": "list_purchase_requests_api_v1_orgs__org_id__purchase_requests_get",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseRequestList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Purchase Requests",
        "tags": [
          "billing"
        ]
      },
      "post": {
        "description": "« Je veux démarrer ». Une demande déjà en cours pour la même offre est renvoyée (200).",
        "operationId": "create_purchase_request_api_v1_orgs__org_id__purchase_requests_post",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseRequestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseRequestOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Purchase Request",
        "tags": [
          "billing"
        ]
      }
    },
...
  }
}